move task components and functions
